Moved combat UI stuff into Combat.tsx, setup combat state tracking in QuestContext
