+++ app/actions/quest.tsx
@@ -11,7 +11,8 @@ import {
 } from './ActionTypes'
 import {SettingsType, XMLElement} from '../reducers/StateTypes'
 import {toCard, toPrevious} from './card'
-import {loadCombatNode, handleAction, CombatResult} from '../parser/Handlers'
+import {handleAction} from '../parser/Handlers'
+import {loadCombatNode, CombatResult} from '../components/Combat'
 import {QuestDetails, QuestContext} from '../reducers/QuestTypes'
 import {ParserNode} from '../parser/Node'
 import {defaultQuestContext} from '../reducers/QuestTypes'
